+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Definitions Change Log
 
-## 2.0.1 under development
+## 2.1.0 October 25, 2022
 
 - Enh #43: Add `Reference::optional()` method that returns `null` when there is no dependency defined
   in container (@vjik)